requirements.txt
huggingface_hub
openai
fastapi
pydantic
load_dotenv
uvicorn